requirements.txt
apispec==5.0.0
attrs==21.2.0
click==8.0.1
dataclasses==0.6
dataclasses-jsonschema==2.14.1
dataclasses-serialization==1.3.1
flasgger==0.9.5
flask==2.0.1
flask-marshmallow==0.14.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
marshmallow==3.13.0
mistune==0.8.4
more-properties==1.1.1
mypy-extensions==0.4.3
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
PyYAML==5.4.1
six==1.16.0
toolz==0.11.1
toposort==1.6
typing-extensions==3.10.0.0
typing-inspect==0.7.1
Werkzeug==2.0.1
marshmallow_dataclass
apifairy